Refactoring kernels
