+++ pypd/integrator.py
@@ -12,7 +12,7 @@ Notes
 """
 
 
-from .kernels.calculate import euler_cromer
+from .kernels.integrator import euler_cromer
 
 
 class Integrator:
