--- pypd/kernels/__init__.py
@@ -1,2 +1,0 @@
-from .calculate import calculate_nodal_forces, calculate_node_damage, smooth_step_data
-from .constitutive_model import linear, trilinear, nonlinear
